requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.7
discord.py==2.3.2
emoji==2.8.0
frozenlist==1.4.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
platformdirs==4.0.0
pylint==3.0.2
python-dotenv==1.0.0
tomli==2.0.1
tomlkit==0.12.2
typing_extensions==4.8.0
yarl==1.9.2
